fix: include provisioner tags when creating template version (#125)
This was unfortunately missed. There's not a great way to assert that
the provisioner tags are even being included in the request. They're
stored in the database for each template version, but we don't expose a
way to retrieve them and verify they were sent.

Treating as hotfix and merging.