fix: use server computed org display name (#183)
I picked up on this reviewing #182, but it's a mistake I made when
reviewing the resource originally 😓

There's a minor inconsistency between display names on orgs and some
other resources with display names on `coderd`. i.e. for organizations:
```
if req.DisplayName == "" {
	req.DisplayName = req.Name
}
```
For some other resources, e.g. users & groups, an empty string display
name is retained, and the CLI/Web UI just shows the actual name if the
display name is empty. For those, it's okay to have an empty string
default in the provider.